+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Bug Fixed
 
+- [x] **$vuepress-plugin-mathjax**: wrong cache for dataurl formula
 - [ ] fix MathJax `Cannot read property 'chars' of undefined`
 
 ### Feature